requirements.txt
python-dotenv==1.0.1
streamlit==1.39.0
google-generativeai